+++ edx_rest_framework_extensions/config.py
@@ -28,9 +28,13 @@ ENABLE_SET_REQUEST_USER_FOR_JWT_COOKIE = 'ENABLE_SET_REQUEST_USER_FOR_JWT_COOKIE
 ENABLE_FORGIVING_JWT_COOKIES = 'ENABLE_FORGIVING_JWT_COOKIES'
 
 # .. setting_name: EDX_DRF_EXTENSIONS[VERIFY_LMS_USER_ID_PROPERTY_NAME]
-# .. setting_default: None
-# .. setting_description: This setting should be set to the name of the user attribute property containing the LMS
-#      user id. Examples might be `id` or `lms_user_id`. If there is no property, leave the default value of None.
-#      This is used by JWT cookie authentication to verify that the (LMS) user id in the JWT is the same
-#      as the LMS user id for a service's session.
+# .. setting_default: None ('lms_user_id' if found)
+# .. setting_description: This setting should be set to the name of the user object property containing the LMS
+#      user id, if one exists. Examples might be 'id' or 'lms_user_id'. To enhance security and provide ease of use
+#      for this setting, if None is supplied, the property 'lms_user_id' will be used if found. In case of unforeseen
+#      issues using lms_user_id, the check can be fully disabled using 'skip-check' as the property name. The default
+#      was not set to 'lms_user_id' directly to avoid misconfiguration logging for services without an lms_user_id
+#      property. The property named by this setting will be used by JWT cookie authentication to verify that the (LMS)
+#      user id in the JWT is the same as the LMS user id for a service's session. This will cause failures in the case
+#      of forgiving cookies, and will simply be used for additional monitoring for successful cookie authentication.
 VERIFY_LMS_USER_ID_PROPERTY_NAME = 'VERIFY_LMS_USER_ID_PROPERTY_NAME'